refactors index.js
